fix(候选A阶段2): 真实数据对拍抓出 1-ULP 浮点结合律 bug
真实 TDX 数据对拍(5股×119bar×3优先级)发现 trades col6(profit) 1-ULP 发散,
合成数据漏掉(真实数据才暴露 commission+stamp_tax 非零):
- 根因: loop.py 预算 comm_stamp=commission+stamp_tax 再 1-comm_stamp = 1-(c+s),
  legacy 是 1-commission-stamp_tax = (1-c)-s, 浮点结合律不同 → 1 ULP
- 修: comm_factor = 1.0 - commission - stamp_tax (同 legacy 结合律)
- 附带: book.get 保留 np.float64 不 cast Python float; ladder_tp 去 float() cast
- 新增 tools/real_parity_check.py: monkeypatch 切壳/legacy, 真实 TDX+QUANTQQ 对拍
- 真实数据 3 优先级全 PASS, 合成 261 passed 零回归

+++ backtest/loop/strategies/ladder_tp.py
@@ -38,16 +38,17 @@ class LadderTpStrategy:
         # mutate bitmask（跨 bar 累计, 与 engine.py:196 一致）
         pos.ladder_done = new_mask
         ratios = ctx.ladder_ratios[: ctx.n_ladder].copy()
-        sell_ratio = float(compute_ladder_sell_ratio(prev_mask, new_mask, profits, ratios))
+        sell_ratio = compute_ladder_sell_ratio(prev_mask, new_mask, profits, ratios)
         # 执行价: 取已置位且 hi_pp 满足的最大档位 profit（engine.py:296-302）
         tp_profit = 0.0
         for li in range(ctx.n_ladder):
             if (new_mask >> li) & 1 and ctx.hi_pp >= profits[li]:
                 if profits[li] > tp_profit:
                     tp_profit = profits[li]
-        exec_price = float(pos.entry_px * (1.0 + tp_profit))
+        # 保留 np.float64 不 cast, 保证与 legacy 裸数组算术字节级一致
+        exec_price = pos.entry_px * (1.0 + tp_profit)
         return [TriggerResult(
             reason=5, strategy_name=self.name, execution_price=exec_price,
             sell_ratio=sell_ratio, is_partial=bool(sell_ratio < 1.0),
-            actual_return=float(tp_profit),
+            actual_return=tp_profit,
         )]
